+++ server/routes/index.js
@@ -2,76 +2,13 @@ const express = require('express')
 const router = express.Router()
 const User = require('../models/User')
 
-const handleLogin = (user, email, password, type) => {
-    if (!user) { 
-        //console.log("user not found case")
-        return 0
-    }
-    else if (user.type != type || user.type != 'Both') { 
-        //console.log("user exists, wrong type case")
-        return 1
-    }
-    else {
-        //console.log("user found, correct type cases")
-        if (password === user.password) { 
-            //console.log("passwords match case")
-            return 2
-        }
-        else { 
-            //console.log("passwords don't match case")
-            return 3
-        }
-    }
-    // Do something on success that logs in as the correct type (customer/owner)
-    // Will need get request back from login page (?)
-}
-
 router.get('/', (req, res) => {
     console.log('absolutely baller man')
 })
 
 router.post('/login', (req, res) => {
-    console.log(req.body.email)
     const email = req.body.email
-    const password = req.body.password
-    const type = req.body.type
-    /*
-    * Check DB for email.
-    *   If email, check type.
-    *       If types match, check for matching password. ENCRYPTION???
-    *           If passwords match, login
-    *           If not, prompt error
-    *       If type is wrong, prompt to confirm Owner/Customer
-    *   If no email, prompt to register
-    * 
-    * WHAT DO YOU DO TO SAVE LOGGED IN STATE?
-    */
-    const caseNum = {num: -1}
-    User.findOne({email: email}).then(user => caseNum.num = handleLogin(user, email, password, type))
-    const caseMsg = {msg: ""}
-    switch (caseNum) {
-        case 0:
-            caseMsg.msg = "Email not registered"
-            console.log(caseMsg.msg)
-            break;
-        case 1:
-            caseMsg.msg = "Email not registered as " + type
-            console.log(caseMsg.msg)
-            break;
-        case 2:
-            caseMsg.msg = "Success"
-            console.log(caseMsg.msg)
-            break;
-        case 3:
-            caseMsg.msg = "Incorrect password"
-            console.log(caseMsg.msg)
-            break;
-        default:
-            console.log("hurr durr")
-            console.log(caseNum)
-    }
-    console.log(caseMsg.msg)
-    res.send(JSON.stringify(caseMsg));
+    User.findOne({email: email}).then(user => res.send(JSON.stringify(user)))
 })
 
 router.post('/register', (req, res) => {
